Add checks for pull request webhooks processors
