+++ app/services/github/webhook_processors/create_pull_request.rb
@@ -9,6 +9,8 @@ module Github
       end
 
       def call
+        return ReturnValue.ok unless DeploymentConfigurationStatus.new(@project).active?(@wrapped_body.full_repo_name)
+
         result = create_project_instance
         create_deployment_links(result)
         ReturnValue.ok
