+++ app/services/github/webhook_processors/update_pull_request.rb
@@ -9,6 +9,8 @@ module Github
       end
 
       def call
+        return ReturnValue.ok unless DeploymentConfigurationStatus.new(@project).active?(@wrapped_body.full_repo_name)
+
         project_instance = @project.project_instances.find_by(attached_pull_request_number: @wrapped_body.number)
         Deployment::Processes::UpdateProjectInstance.new(project_instance, get_user(@wrapped_body.sender)).call
         ReturnValue.ok
