+++ app/views/layouts/application.html.slim
@@ -14,7 +14,7 @@ html
     link href="https://fonts.googleapis.com/css?family=Titillium+Web:300,300i,400,400i,700,700i&display=swap" rel="stylesheet"
 
     = stylesheet_pack_tag 'application', media: 'all', 'data-turbolinks-track': 'reload'
-    = javascript_pack_tag 'application', media: 'all', 'data-turbolinks-track': 'reload'
+    = javascript_pack_tag 'application', media: 'all', 'data-turbolinks-track': 'reload', async: true
 
   body
     .body-wrapper
